Moved all php functions into a new Folder "includes" and added a "require_once" at top of the php code.
